Add api.query
